# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Cafe Manager API",
    "version": "1.0.0",
    "description": "REST API for the Smart Cafe Manager multi-tenant platform — Next.js App Router Route Handlers backed by Supabase. This spec is generated from `@swagger` JSDoc blocks on route handlers by scripts/generateSwagger.js; view it rendered at /api-docs. Only a representative subset of routes carries `@swagger` annotations so far (see the plan doc, Phase 11) — extend the pattern to the rest incrementally by adding the same style of block above any other exported GET/POST/PATCH/DELETE handler."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this app"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by /api/auth/staff-login, /api/auth/customer-login, or /api/auth/register. Send as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "Bill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "siteId": {
            "type": "string",
            "nullable": true
          },
          "spaceId": {
            "type": "string",
            "nullable": true
          },
          "customerId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "paid",
              "company_charged"
            ]
          },
          "paymentMethod": {
            "type": "string",
            "nullable": true,
            "enum": [
              "cash",
              "online",
              null
            ]
          },
          "totalAmount": {
            "type": "number"
          },
          "couponCodeId": {
            "type": "string",
            "nullable": true
          },
          "razorpayOrderId": {
            "type": "string",
            "nullable": true
          },
          "cashCollectedBy": {
            "type": "string",
            "nullable": true
          },
          "cashCollectedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "orderIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/bills/{id}/pay/razorpay": {
      "post": {
        "tags": [
          "Bills"
        ],
        "summary": "Create a Razorpay order for a pending bill",
        "description": "Creates the Razorpay order for the bill (possibly coupon-discounted) total and returns the details the client checkout widget needs. The bill is only marked paid once the /api/webhooks/razorpay/bill-payment webhook confirms payment. Requires a bearer token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Razorpay order created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "razorpayOrderId": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "integer",
                      "description": "Amount in paise"
                    },
                    "currency": {
                      "type": "string",
                      "example": "INR"
                    },
                    "keyId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bill is already settled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Bill not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bills/{id}/pay/cash": {
      "post": {
        "tags": [
          "Bills"
        ],
        "summary": "Record the guest choice to pay a bill with cash",
        "description": "Sets payment_method to cash on a pending bill so staff see it in their pending-cash-bills view. Does not mark the bill paid — see /api/bills/{id}/collect-cash for that. Requires a bearer token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cash payment choice recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bill"
                }
              }
            }
          },
          "400": {
            "description": "Bill is already settled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Bill not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bills/{id}/collect-cash": {
      "post": {
        "tags": [
          "Bills"
        ],
        "summary": "Confirm cash has been physically collected for a bill",
        "description": "Owner or Manager only. Marks a pending cash-payment-method bill as paid. Safe to call twice - the second call matches zero rows and 404s instead of double-processing.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cash collection confirmed, bill marked paid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bill"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an Owner or Manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No pending cash bill found with that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bills/generate": {
      "post": {
        "tags": [
          "Bills"
        ],
        "summary": "Aggregate unbilled orders into a new bill",
        "description": "Customer-initiated bills their own unbilled orders. Owner/Manager may pass spaceId to bill an entire table/room, across whichever guest accounts ordered from it. Requires a bearer token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "spaceId": {
                    "type": "string",
                    "description": "Owner/Manager only: bill every unbilled order from this space"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bill created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bill"
                }
              }
            }
          },
          "400": {
            "description": "No unbilled orders found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}